Extract WorkspaceGate and PWAInstallPrompt from layout.tsx
Move WorkspaceGate (153 lines) and PWAInstallPrompt (49 lines)
from the dashboard layout into their own component files. The
layout file shrinks from 264 lines to 55 lines.
